Refactoring code

+++ server_flask/app/handlers/handler.py
@@ -1,3 +1,4 @@
+import imghdr
 import os
 import re
 
@@ -7,7 +8,7 @@ from pydantic import ValidationError
 from sqlalchemy import desc, select
 
 from ..depends.depend import current_user
-from ..model.models import AnketaSchemaJson
+from ..model.models import AnketaSchemaJson, models_tables
 from ..model.tables import Users, db_session, Persons, tables_models
 
 
@@ -41,6 +42,29 @@ def handle_get_item(item, item_id):
     return result[0] if item == "persons" else result
 
 
+def handle_post_item(json_data, item, item_id=""):
+    """
+    Updates an item in the database based on the provided JSON data, item, and item_id.
+
+    Args:
+        json_data (dict): A dictionary containing the data to update the item.
+        item (str): The type of item to update in the database.
+        item_id (int): The ID of the item to update.
+
+    Returns:
+        None
+    """
+    table, model = tables_models.get(item), models_tables.get(item)
+    if not model or not table:
+        return abort(400)
+    json_dict = model(**json_data).dict()
+    if item != "persons":
+        json_dict["person_id"] = item_id
+    json_dict["user_id"] = current_user.get("id")
+    db_session.merge(model(**json_dict))
+    db_session.commit()
+
+
 def handle_post_resume(resume):
     """
     Updates a resume in the database with the provided data.
@@ -58,75 +82,50 @@ def handle_post_resume(resume):
     if not re.match(r"[А-ЯЁЙа-яё]", resume["surname"][0]):
         return abort(400)
     resume["editable"] = True
-    resume["user_id"] = current_user["id"]
-    resume["region"] = current_user["region"]
-    if not resume.get("id"):
-        person = db_session.execute(
-            select(Persons).where(
-                Persons.surname.ilike("%{}%".format(resume["surname"])),
-                Persons.firstname.ilike("%{}%".format(resume["firstname"])),
-                Persons.patronymic.ilike("%{}%".format(resume["patronymic"])),
-                Persons.birthday == resume["birthday"],
-            )
-        ).scalar_one_or_none()
-        if not person:
-            person = Persons(**resume)
-            db_session.add(person)
-            db_session.flush()
-            person.destination = make_destination(
-                resume["region"],
-                resume["surname"],
-                resume["firstname"],
-                resume.get("patronymic", ""),
-                person.id,
-            )
-            db_session.commit()
-            return [person.id, person.destination]
-        else:
-            if not person.destination:
-                person.destination = make_destination(
-                    resume["region"],
-                    resume["surname"],
-                    resume["firstname"],
-                    resume.get("patronymic", ""),
-                    person.id,
-                )
-                db_session.commit()
-            if person.editable:
-                return [person.id, person.destination]
-            resume["id"] = person.id
-            handle_post_item(resume, "persons")
-            return [resume["id"], person.destination]
+    resume["user_id"] = current_user.get("id")
+    resume["region"] = current_user.get("region")
+    person = db_session.execute(
+        select(Persons).where(
+            Persons.surname.ilike("%{}%".format(resume["surname"])),
+            Persons.firstname.ilike("%{}%".format(resume["firstname"])),
+            Persons.patronymic.ilike("%{}%".format(resume["patronymic"])),
+            Persons.birthday == resume["birthday"],
+        )
+    ).scalar_one_or_none()
+    if not person:
+        person = Persons(**resume)
+        db_session.add(person)
+        db_session.flush()
+        person.destination = make_destination(
+            resume["region"],
+            resume["surname"],
+            resume["firstname"],
+            resume.get("patronymic", ""),
+            person.id,
+        )
+        db_session.commit()
+        return person.id
+    if person.editable:
+        return person.id
+    if not person.destination:
+        person.destination = make_destination(
+            resume["region"],
+            resume["surname"],
+            resume["firstname"],
+            resume.get("patronymic", ""),
+            person.id,
+        )
+        db_session.commit()
+    resume["id"] = person.id
     handle_post_item(resume, "persons")
-
-
-def handle_post_item(json_dict, item, item_id=""):
-    """
-    Updates an item in the database based on the provided JSON data, item, and item_id.
-
-    Args:
-        json_data (dict): A dictionary containing the data to update the item.
-        item (str): The type of item to update in the database.
-        item_id (int): The ID of the item to update.
-
-    Returns:
-        None
-    """
-    model = tables_models.get(item)
-    if not model:
-        return abort(400)
-    if item != "persons":
-        json_dict["person_id"] = item_id
-    json_dict["user_id"] = current_user["id"]
-    db_session.merge(model(**json_dict))
-    db_session.commit()
+    return resume["id"]
 
 
 def handle_json_to_dict(data):
     try:
         anketa = AnketaSchemaJson(**data).dict()
         anketa["resume"] = {
-            "region": current_user["region"],
+            "region": current_user.get("region"),
             "surname": anketa.pop("surname", "").upper().strip(),
             "firstname": anketa.pop("firstname", "").upper().strip(),
             "patronymic": anketa.pop("patronymic", "").upper().strip()
@@ -192,12 +191,13 @@ def handle_image(file, item_dir):
     Returns:
         None
     """
-    image = Image.open(file)
-    image = image.convert("RGB")
-    new_file = os.path.join(item_dir, "image.jpg")
-    if os.path.isfile(new_file):
-        os.remove(new_file)
-    image.save(new_file, format="JPEG", quality=90)
+    if imghdr.what(file) is not None:
+        image = Image.open(file)
+        image = image.convert("RGB")
+        new_file = os.path.join(item_dir, "image.jpg")
+        if os.path.isfile(new_file):
+            os.remove(new_file)
+        image.save(new_file, format="JPEG", quality=90)
 
 
 def make_destination(region, surname, firstname, patronymic, person_id):
